ユーザー削除、チーム編集ボタンの条件分岐追加

+++ app/controllers/assigns_controller.rb
@@ -27,15 +27,21 @@ class AssignsController < ApplicationController
   end
 
   def assign_destroy(assign, assigned_user)
-    if assigned_user == assign.team.owner
-      I18n.t('views.messages.cannot_delete_the_leader')
-    elsif Assign.where(user_id: assigned_user.id).count == 1
-      I18n.t('views.messages.cannot_delete_only_a_member')
-    elsif assign.destroy
-      set_next_team(assign, assigned_user)
-      I18n.t('views.messages.delete_member')
+    if current_user == assign.team.owner || assign.user == current_user
+      if assigned_user == assign.team.owner
+        I18n.t('views.messages.cannot_delete_the_leader')
+      elsif assign.user != current_user
+        I18n.t('views.messages.cannot_delete_other_users')
+      elsif Assign.where(user_id: assigned_user.id).count == 1
+        I18n.t('views.messages.cannot_delete_only_a_member')
+      elsif assign.destroy
+        set_next_team(assign, assigned_user)
+        I18n.t('views.messages.delete_member')
+      else
+        I18n.t('views.messages.cannot_delete_member_4_some_reason')
+      end
     else
-      I18n.t('views.messages.cannot_delete_member_4_some_reason')
+      I18n.t('views.messages.cannot_delete_users')
     end
   end
 
@@ -45,23 +51,19 @@ class AssignsController < ApplicationController
       redirect_to team_url(team), notice: I18n.t('views.messages.email_already_exists')
     end
   end
-
   def email_reliable?(address)
     address.match(/\A[\w+\-.]+@[a-z\d\-.]+\.[a-z]+\z/i)
   end
-
   def user_exist?
     team = find_team(params[:team_id])
     unless User.exists?(email: params[:email])
       redirect_to team_url(team), notice: I18n.t('views.messages.does_not_exist_email')
     end
   end
-
   def set_next_team(assign, assigned_user)
     another_team = Assign.find_by(user_id: assigned_user.id).team
     change_keep_team(assigned_user, another_team) if assigned_user.keep_team_id == assign.team_id
   end
-
   def find_team(team_id)
     Team.friendly.find(params[:team_id])
   end
